[FSSDK-12368] Address PR review: simplify types, relocate tests to module-level files
- Remove NotRequired import and use Optional[list[str]] for HoldoutDict.includedRules
- Migrate test fixtures from includedFlags/excludedFlags to includedRules
- Move LocalHoldoutsTest into test_config.py (HoldoutConfigTest + LocalHoldoutConfigTest)
- Move LocalHoldoutsDecisionFlowTest into test_decision_service_holdout.py
- Delete test_local_holdouts.py (violated module-level test convention)
- Tighten weakened assertions to exact counts and content checks
