+++ optimizely/helpers/types.py
@@ -21,11 +21,6 @@ if version_info < (3, 8):
 else:
     from typing import TypedDict  # type: ignore
 
-if version_info < (3, 11):
-    from typing_extensions import NotRequired
-else:
-    from typing import NotRequired  # type: ignore
-
 
 # Intermediate types for type checking deserialized datafile json before actual class instantiation.
 # These aren't used for anything other than type signatures
@@ -135,4 +130,4 @@ class HoldoutDict(ExperimentDict):
     Extends ExperimentDict with holdout-specific properties.
     """
     holdoutStatus: HoldoutStatus
-    includedRules: NotRequired[Optional[list[str]]]
+    includedRules: Optional[list[str]]
